add millisecond and microsecond format strings
It's important when looking at time-ordered logs to have resolution
below a second. I think this benefit outweighs being strictly
POSIX-compliant. This change allows for the format specifiers `%L` for
milliseconds and `%f` for microseconds, which are the format strings
used in the extensions provided by the github.com/lestrrat-go/strftime
library.